Bugfix: Do not send multiple responses.
When try to restrict the GET Method you will get an Node.js error:
```Can't set headers after they are sent to the client```.
After failing the acl check it will still execute the usual context which results in sending an additional answer.  A simple check if the response is already finished solves this problem.